+++ gui/src/App.tsx
@@ -14,6 +14,7 @@ import { ProviderWallets } from "./components/ProviderWallets";
 import { TxReviewDialog } from "./components/TxReviewDialog";
 import { WagmiWrapper } from "./components/WagmiWrapper";
 import { WalletUnlockDialog } from "./components/WalletUnlockDialog";
+import { useStore } from "./store";
 
 const queryClient = new QueryClient({
   defaultOptions: { queries: { suspense: true } },
@@ -25,40 +26,38 @@ export default function App() {
       <ProviderTheme>
         <CssBaseline>
           <QueryClientProvider client={queryClient}>
-            <ProviderWallets>
-              <ProviderNetworks>
-                <WagmiWrapper>
-                  <ProviderCurrentNetwork>
-                    <ProviderNativeBalance>
-                      <ProviderTokensBalances>
-                        <Router>
-                          <Switch>
-                            <Route path="/dialog/tx-review/:id">
-                              {({ id }: { id: string }) => (
-                                <TxReviewDialog id={parseInt(id)} />
-                              )}
-                            </Route>
+            <ProviderNetworks>
+              <WagmiWrapper>
+                <ProviderCurrentNetwork>
+                  <ProviderNativeBalance>
+                    <ProviderTokensBalances>
+                      <Router>
+                        <Switch>
+                          <Route path="/dialog/tx-review/:id">
+                            {({ id }: { id: string }) => (
+                              <TxReviewDialog id={parseInt(id)} />
+                            )}
+                          </Route>
 
-                            <Route path="/dialog/wallet-unlock/:id">
-                              {({ id }: { id: string }) => (
-                                <WalletUnlockDialog id={parseInt(id)} />
-                              )}
-                            </Route>
+                          <Route path="/dialog/wallet-unlock/:id">
+                            {({ id }: { id: string }) => (
+                              <WalletUnlockDialog id={parseInt(id)} />
+                            )}
+                          </Route>
 
-                            <Route>
-                              <CommandBar>
-                                <Navbar />
-                                <HomePage />
-                              </CommandBar>
-                            </Route>
-                          </Switch>
-                        </Router>
-                      </ProviderTokensBalances>
-                    </ProviderNativeBalance>
-                  </ProviderCurrentNetwork>
-                </WagmiWrapper>
-              </ProviderNetworks>
-            </ProviderWallets>
+                          <Route>
+                            <CommandBar>
+                              <Navbar />
+                              <HomePage />
+                            </CommandBar>
+                          </Route>
+                        </Switch>
+                      </Router>
+                    </ProviderTokensBalances>
+                  </ProviderNativeBalance>
+                </ProviderCurrentNetwork>
+              </WagmiWrapper>
+            </ProviderNetworks>
           </QueryClientProvider>
         </CssBaseline>
       </ProviderTheme>
