--- gui/src/hooks/useWallets.tsx
@@ -1,7 +1,0 @@
-import { useContext } from "react";
-
-import { WalletsContext } from "../components/ProviderWallets";
-
-export function useWallets() {
-  return useContext(WalletsContext);
-}
